refactor: AdminLoginPage, MainPage 컴포넌트 정리
