+++ src/components/common/LazyComponents.jsx
@@ -1,8 +1,8 @@
 import { lazy } from 'react';
 
 // 페이지 컴포넌트들 지연 로딩
-export const LazyMainPage = lazy(() => import('../pages/Main/MainPage'));
-export const LazyAdminLoginPage = lazy(() => import('../pages/AdminLogin/AdminLoginPage'));
+export const LazyMainPage = lazy(() => import('../../pages/MainPage'));
+export const LazyAdminLoginPage = lazy(() => import('../../pages/AdminLoginPage'));
 export const LazyRecruitingManagePage = lazy(() => import('../../pages/RecruitingManagePage'));
 export const LazyRecruitingDetailPage = lazy(() => import('../../pages/RecruitingDetailPage'));
 
@@ -10,4 +10,4 @@ export const LazyRecruitingDetailPage = lazy(() => import('../../pages/Recruitin
 export const LazyCreateRecruitingModal = lazy(() => import('../../features/admin/CreateRecruitingModal'));
 export const LazyAdminCodeModal = lazy(() => import('../../features/admin/AdminCodeModal'));
 export const LazyAdminManageModal = lazy(() => import('../../features/admin/AdminManageModal'));
-export const LazyEmailModal = lazy(() => import('../../features/recruiting/EmailModal'));
+export const LazyEmailModal = lazy(() => import('../recruiting/modals/EmailModal'));
