--- src/components/pages/AdminLogin/components/index.js
@@ -1,1 +1,0 @@
-export { default as LoginForm } from './LoginForm';
