+++ src/components/recruiting/detail/ApplicantListSection.css
