+++ src/components/recruiting/detail/RecruitingInfoSection.jsx
