+++ src/components/recruiting/detail/index.js
@@ -1,0 +1,4 @@
+export { default as ApplicantListSection } from './ApplicantListSection';
+export { default as RecruitingHeader } from './RecruitingHeader';
+export { default as RecruitingInfoSection } from './RecruitingInfoSection';
+export { default as ApplicantCard } from './ApplicantCard';
