+++ src/components/recruiting/evaluation/index.js
@@ -1,0 +1,1 @@
+export { default as EvaluationForm } from './EvaluationForm';
