+++ src/components/recruiting/filters/index.js
@@ -1,0 +1,1 @@
+export { default as ApplicantFilters } from './ApplicantFilters';
