+++ src/components/recruiting/index.js
@@ -1,0 +1,14 @@
+// Detail components
+export * from './detail';
+
+// Modal components  
+export * from './modals';
+
+// Stats components
+export * from './stats';
+
+// Evaluation components
+export * from './evaluation';
+
+// Filter components
+export * from './filters';
