+++ src/components/recruiting/modals/EmailModal.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Mail, Send, Users as UsersIcon, Type, MessageCircle } from 'lucide-react';
-import '../../components/common/CommonModals.css';
+import '../../common/CommonModals.css';
 
 const EmailModal = ({ 
   isOpen,
