+++ src/components/recruiting/modals/index.js
@@ -1,0 +1,3 @@
+export { default as BulkStatusChangeModal } from './BulkStatusChangeModal';
+export { default as ApplicantModal } from './ApplicantModal';
+export { default as EmailModal } from './EmailModal';
