+++ src/components/recruiting/stats/index.js
@@ -1,0 +1,1 @@
+export { default as StatsSection } from './StatsSection';
